# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "포포춘(For Fortune) 운세 API",
    "version": "2.0.0",
    "description": "AI 기반 운세 상담 플랫폼 - 사주, 타로, 꿈해몽, 행운번호 등"
  },
  "servers": [
    {
      "url": "http://localhost:3350",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 액세스 토큰을 Bearer 형식으로 전달"
      }
    },
...
  },
  "paths": {
...
    "/api/auth/profile": {
      "get": {
        "operationId": "getUserProfile",
        "summary": "사용자 프로필 조회",
        "description": "현재 로그인한 사용자의 정보를 조회합니다.",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "프로필 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "nullable": true,
                          "description": "사용자 이메일"
                        },
                        "phone": {
                          "type": "string",
                          "nullable": true,
                          "description": "사용자 연락처"
                        },
                        "chatUsableUntil": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true,
                          "description": "계정 기준 채팅 이용 가능 만료 시각"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "USER",
                            "ADMIN"
                          ],
                          "description": "사용자 역할"
                        },
                        "termsAgreed": {
                          "type": "boolean",
                          "description": "이용약관 동의 여부"
                        },
                        "privacyAgreed": {
                          "type": "boolean",
                          "description": "개인정보 수집·이용 동의 여부"
                        },
                        "marketingAgreed": {
                          "type": "boolean",
                          "description": "마케팅 정보 수신 동의 여부"
                        },
                        "authType": {
                          "type": "string",
                          "description": "주요 소셜 인증 타입"
                        },
                        "socialRelinkRequired": {
                          "type": "boolean",
                          "description": "소셜 계정 재연동 필요 여부"
                        },
                        "socialRelinkMessage": {
                          "type": "string",
                          "nullable": true,
                          "description": "재연동 안내 메시지"
                        },
                        "socialAccounts": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "provider": {
                                "type": "string"
                              },
                              "providerUid": {
                                "type": "string"
                              },
                              "connectedAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "사용자 정보를 조회했습니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요"
          },
          "404": {
            "description": "사용자를 찾을 수 없음"
          }
        }
      },
      "patch": {
        "operationId": "updateUserProfile",
        "summary": "사용자 프로필 업데이트",
        "description": "현재 로그인한 사용자의 정보를 부분 업데이트합니다. 값이 있는 필드만 업데이트됩니다.",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string",
                    "description": "닉네임",
                    "example": "홍길동"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "이메일 주소",
                    "example": "user@example.com"
                  },
                  "phone": {
                    "type": "string",
                    "description": "연락처",
                    "example": "010-1234-5678"
                  },
                  "termsAgreed": {
                    "type": "boolean",
                    "description": "이용약관 동의 여부"
                  },
                  "privacyAgreed": {
                    "type": "boolean",
                    "description": "개인정보 수집·이용 동의 여부"
                  },
                  "marketingAgreed": {
                    "type": "boolean",
                    "description": "마케팅 정보 수신 동의 여부"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "프로필 업데이트 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "nullable": true
                        },
                        "phone": {
                          "type": "string",
                          "nullable": true
                        },
                        "chatUsableUntil": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "USER",
                            "ADMIN"
                          ]
                        },
                        "termsAgreed": {
                          "type": "boolean"
                        },
                        "privacyAgreed": {
                          "type": "boolean"
                        },
                        "marketingAgreed": {
                          "type": "boolean"
                        },
                        "authType": {
                          "type": "string",
                          "nullable": true
                        },
                        "socialAccounts": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "사용자 정보가 업데이트되었습니다."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (이메일 중복, 업데이트할 정보 없음 등)"
          },
          "401": {
            "description": "인증 필요"
          },
          "404": {
            "description": "사용자를 찾을 수 없음"
          }
        }
      }
    },
...
    "/api/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "로그아웃",
        "description": "현재 세션을 종료하고 토큰을 무효화합니다.",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "로그아웃 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "로그아웃이 완료되었습니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요"
          }
        }
      }
    },
...
    "/api/auth/account": {
      "delete": {
        "operationId": "deleteAccount",
        "summary": "계정 탈퇴",
        "description": "사용자 계정을 완전히 삭제합니다.",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "계정 삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "계정이 삭제되었습니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
    "/api/auth/{provider}": {
      "get": {
        "operationId": "startSocialLogin",
        "summary": "소셜 로그인 시작",
        "description": "카카오, 구글, 네이버 소셜 로그인을 시작합니다.",
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "kakao",
                "google",
                "naver"
              ]
            },
            "description": "소셜 로그인 제공자"
          }
        ],
        "responses": {
          "302": {
            "description": "소셜 로그인 페이지로 리다이렉트"
          }
        }
      }
    },
...
  }
}